refactor: extract shared components for Evals pages
- Extract PageHeader, LoadingState, AgentVersionSelect components
- Refactor ab-tests/page.tsx into smaller modules (215 lines)
- Update LLM-as-Judge and Head-to-Head pages to use shared components
