+++ admin-next/src/app/(dashboard)/evals/components/loading-state.tsx
@@ -1,0 +1,11 @@
+interface LoadingStateProps {
+  message?: string;
+}
+
+export function LoadingState({ message = 'Loading...' }: LoadingStateProps) {
+  return (
+    <div className="flex items-center justify-center h-64">
+      <div className="text-neutral-400">{message}</div>
+    </div>
+  );
+}
