+++ admin-next/src/app/(dashboard)/evals/components/page-header.tsx
@@ -1,0 +1,17 @@
+interface PageHeaderProps {
+  title: string;
+  description: string;
+  action?: React.ReactNode;
+}
+
+export function PageHeader({ title, description, action }: PageHeaderProps) {
+  return (
+    <header className="mb-6 flex items-center justify-between">
+      <div>
+        <h1 className="text-2xl font-bold text-white">{title}</h1>
+        <p className="mt-1 text-sm text-neutral-400">{description}</p>
+      </div>
+      {action}
+    </header>
+  );
+}
